refactored everything to use RequestBody, added new controller method…

+++ src/main/java/com/revature/hydra/skilltype/service/SkillTypeCompositionService.java
@@ -1,46 +1,92 @@
 package com.revature.hydra.skilltype.service;
 
+import java.util.List;
+
 import org.springframework.amqp.core.AmqpTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.revature.beans.SimpleSkillType;
+import com.revature.beans.SimpleSkillTypeBucketLookup;
 import com.revature.hydra.skilltype.data.SimpleSkillTypeBucketLookupRepository;
 import com.revature.hydra.skilltype.data.SimpleSkillTypeRepository;
 
 @Service
 public class SkillTypeCompositionService {
-	
+
 	@Autowired
 	AmqpTemplate rabbitTemplate;
-	
+
 	@Autowired
 	SimpleSkillTypeRepository skillTypeRepository;
-	
+
 	@Autowired
 	SimpleSkillTypeBucketLookupRepository skillTypeBucketLookupRepository;
-	
-	public void createSkill(Integer skillTypeId, String skillTypeName, Integer[] bucketIds, Integer[] weights) {
-		SimpleSkillType st = new SimpleSkillType(skillTypeId, skillTypeName, true);
-		skillTypeRepository.save(st);
-		for(int i = 0; i < bucketIds.length;i++) {
-			/**
-			 * SimpleBucket sb = (SimpleBucket)bucketReceiver.find(bucketIds[i]);
-			 * SimpleSkillTypeBucketLookup stbl = new SimpleSkillTypeBucketLookup(123, st, sb, weights[i];
-			 */
-		}
+
+	public SimpleSkillType findSkillTypeByName(String name) {
+		SimpleSkillType skillType = skillTypeRepository.findBySkillTypeName(name);
+		return skillType;
 	}
-	
-	public void updateSkill(Integer skillTypeId, String skillTypeName, Integer[] bucketIds, Integer[] weights) {
-		SimpleSkillType st = skillTypeRepository.findOne(skillTypeId);
-		st.setSkillTypeName(skillTypeName);
+
+	public void createSkill(String skillTypeName) {
+		SimpleSkillType st = new SimpleSkillType(skillTypeName, true);
 		skillTypeRepository.save(st);
-		skillTypeBucketLookupRepository.deleteBySkillType(st);
-		for(int i = 0; i < bucketIds.length; i++) {
-			/**
-			 * SimpleBucket sb = (SimpleBucket)bucketReceiver.find(bucketIds[i]);
-			 * SimpleSkillBucketLookup stbl = new SimpleSkillTypeBucketLookup(123, st, sb, weights[i]);
-			 */
+	}
+
+	/**
+	 * Create a new association between one skillType and multiple buckets.
+	 * 
+	 * Checks to see if the weights add up to 100% before performing further
+	 * actions. Deletes skillType from table before adding the information.
+	 * 
+	 * @param skillTypeId
+	 * @param skillTypeName
+	 * @param bucketIds
+	 * @param weights
+	 */
+	public Boolean updateSkillTypeBucketLookup(Integer skillTypeId, String skillTypeName, Integer[] bucketIds,
+			Double[] weights) {
+		Double sum = 0.0;
+		for (Double x : weights) {
+			sum += x;
 		}
+
+		if (sum == 100d) {
+			SimpleSkillType st = skillTypeRepository.findOne(skillTypeId);
+			List<SimpleSkillTypeBucketLookup> sstbl = skillTypeBucketLookupRepository.findAllBySkillType(st);
+			
+			// Update skillType. 
+			// Delete old skillType Bucket associations
+			if (sstbl.size() == 0 && st != null) {
+				skillTypeBucketLookupRepository.deleteBySkillType(st);
+				
+				for (int i = 0; i < bucketIds.length && i < weights.length; i++) {
+					skillTypeBucketLookupRepository.save(new SimpleSkillTypeBucketLookup(st, bucketIds[i], weights[i]));
+				}
+				return true;
+			}
+		}
+		return false;
+	}
+
+	public Boolean createSkillTypeBucketLookup(String skillTypeName, Integer[] bucketIds, Double[] weights) {
+		Double sum = 0.0;
+		for (Double x : weights) {
+			sum += x;
+		}
+
+		if (sum == 100d) {
+			SimpleSkillType st = skillTypeRepository.findBySkillTypeName(skillTypeName);
+			List<SimpleSkillTypeBucketLookup> sstbl = skillTypeBucketLookupRepository.findAllBySkillType(st);
+			// Create a new skill type association if it doesn't exist.
+			if (sstbl.size() == 0 && st != null) {
+				// Go through both arrays and add entries accordingly.
+				for (int i = 0; i < bucketIds.length && i < weights.length; i++) {
+					skillTypeBucketLookupRepository.save(new SimpleSkillTypeBucketLookup(st, bucketIds[i], weights[i]));
+				}
+				return true;
+			}
+		}
+		return false;
 	}
 }
